Update version 2.4
Game state machine logic is moved from task game to a separate functions in ReactionGame.c and the respective task is moved to main.c, from where it calls the function.
The task tft is also moved to main.c.
TODO: Check the reaction time capture functionality. It seems that the total time calculated is wrong.